@@ -1,34 +1,20 @@
 package queue
 
 import (
-	"fmt"
+	"github.com/ekristen/azure-nuke/pkg/nuke"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/queue"
 
 	"github.com/ekristen/azure-nuke/pkg/azure"
-	"github.com/ekristen/azure-nuke/pkg/log"
-	"github.com/ekristen/azure-nuke/pkg/resource"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/log"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/resource"
 
 	_ "github.com/ekristen/azure-nuke/resources"
-)
-
-type ItemState int
-
-// States of Items based on the latest request to AWS.
-const (
-	ItemStateNew ItemState = iota
-	ItemStatePending
-	ItemStateWaiting
-	ItemStateFailed
-	ItemStateFiltered
-	ItemStateFinished
 )
 
 // An Item describes an actual AWS resource entity with the current state and
 // some metadata.
 type Item struct {
-	Resource resource.Resource
-
-	State  ItemState
-	Reason string
+	queue.Item
 
 	Authorizers    azure.Authorizers
 	TenantId       string
@@ -41,24 +27,24 @@ type Item struct {
 
 func (i *Item) Print() {
 	switch i.State {
-	case ItemStateNew:
+	case queue.ItemStateNew:
 		log.Log(i.SubscriptionId, i.Type, i.Resource, log.ReasonWaitPending, "would remove")
-	case ItemStatePending:
+	case queue.ItemStatePending:
 		log.Log(i.SubscriptionId, i.Type, i.Resource, log.ReasonWaitPending, "triggered remove")
-	case ItemStateWaiting:
+	case queue.ItemStateWaiting:
 		log.Log(i.SubscriptionId, i.Type, i.Resource, log.ReasonWaitPending, "waiting")
-	case ItemStateFailed:
+	case queue.ItemStateFailed:
 		log.Log(i.SubscriptionId, i.Type, i.Resource, log.ReasonError, "failed")
-	case ItemStateFiltered:
+	case queue.ItemStateFiltered:
 		log.Log(i.SubscriptionId, i.Type, i.Resource, log.ReasonSkip, i.Reason)
-	case ItemStateFinished:
+	case queue.ItemStateFinished:
 		log.Log(i.SubscriptionId, i.Type, i.Resource, log.ReasonSuccess, "removed")
 	}
 }
 
 // List gets all resource items of the same resource type like the Item.
 func (i *Item) List() ([]resource.Resource, error) {
-	listers := resource.GetListersV2()
+	listers := resource.GetListers()
 	/*
 		sess, err := i.Region.Session(i.Type)
 		if err != nil {
@@ -66,57 +52,12 @@ func (i *Item) List() ([]resource.Resource, error) {
 		}
 		return listers[i.Type](sess)
 	*/
-	return listers[i.Type](resource.ListerOpts{
+	return listers[i.Type](nuke.ListerOpts{
 		Authorizers:    i.Authorizers,
 		TenantId:       i.TenantId,
 		SubscriptionId: i.SubscriptionId,
 		ResourceGroup:  i.ResourceGroup,
 	})
-}
-
-func (i *Item) GetProperty(key string) (string, error) {
-	if key == "" {
-		stringer, ok := i.Resource.(resource.LegacyStringer)
-		if !ok {
-			return "", fmt.Errorf("%T does not support legacy IDs", i.Resource)
-		}
-		return stringer.String(), nil
-	}
-
-	getter, ok := i.Resource.(resource.ResourcePropertyGetter)
-	if !ok {
-		return "", fmt.Errorf("%T does not support custom properties", i.Resource)
-	}
-
-	return getter.Properties().Get(key), nil
-}
-
-func (i *Item) Equals(o resource.Resource) bool {
-	iType := fmt.Sprintf("%T", i.Resource)
-	oType := fmt.Sprintf("%T", o)
-	if iType != oType {
-		return false
-	}
-
-	iStringer, iOK := i.Resource.(resource.LegacyStringer)
-	oStringer, oOK := o.(resource.LegacyStringer)
-	if iOK != oOK {
-		return false
-	}
-	if iOK && oOK {
-		return iStringer.String() == oStringer.String()
-	}
-
-	iGetter, iOK := i.Resource.(resource.ResourcePropertyGetter)
-	oGetter, oOK := o.(resource.ResourcePropertyGetter)
-	if iOK != oOK {
-		return false
-	}
-	if iOK && oOK {
-		return iGetter.Properties().Equals(oGetter.Properties())
-	}
-
-	return false
 }
 
 type Queue []*Item
@@ -125,7 +66,7 @@ func (q Queue) CountTotal() int {
 	return len(q)
 }
 
-func (q Queue) Count(states ...ItemState) int {
+func (q Queue) Count(states ...queue.ItemState) int {
 	count := 0
 	for _, item := range q {
 		for _, state := range states {
