@@ -2,35 +2,38 @@ package resources
 
 import (
 	"context"
+	"github.com/ekristen/azure-nuke/pkg/nuke"
 	"time"
 
 	"github.com/sirupsen/logrus"
 
 	"github.com/Azure/azure-sdk-for-go/services/web/mgmt/2021-03-01/web"
 
-	"github.com/ekristen/azure-nuke/pkg/resource"
-	"github.com/ekristen/azure-nuke/pkg/types"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/resource"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/types"
 )
 
-type AppServicePlan struct {
-	client web.AppServicePlansClient
-	name   string
-	rg     string
-}
-
 func init() {
-	resource.RegisterV2(resource.Registration{
+	resource.Register(resource.Registration{
 		Name:   "AppServicePlan",
-		Scope:  resource.ResourceGroup,
-		Lister: ListAppServicePlan,
+		Scope:  nuke.ResourceGroup,
+		Lister: AppServicePlanLister{},
 	})
 }
 
-func ListAppServicePlan(opts resource.ListerOpts) ([]resource.Resource, error) {
-	logrus.Tracef("subscription id: %s", opts.SubscriptionId)
+type AppServicePlanLister struct {
+	opts nuke.ListerOpts
+}
 
-	client := web.NewAppServicePlansClient(opts.SubscriptionId)
-	client.Authorizer = opts.Authorizers.Management
+func (l AppServicePlanLister) SetOptions(opts interface{}) {
+	l.opts = opts.(nuke.ListerOpts)
+}
+
+func (l AppServicePlanLister) List() ([]resource.Resource, error) {
+	logrus.Tracef("subscription id: %s", l.opts.SubscriptionId)
+
+	client := web.NewAppServicePlansClient(l.opts.SubscriptionId)
+	client.Authorizer = l.opts.Authorizers.Management
 	client.RetryAttempts = 1
 	client.RetryDuration = time.Second * 2
 
@@ -39,7 +42,7 @@ func ListAppServicePlan(opts resource.ListerOpts) ([]resource.Resource, error) {
 	logrus.Trace("attempting to list ssh key")
 
 	ctx := context.Background()
-	list, err := client.ListByResourceGroup(ctx, opts.ResourceGroup)
+	list, err := client.ListByResourceGroup(ctx, l.opts.ResourceGroup)
 	if err != nil {
 		return nil, err
 	}
@@ -52,7 +55,7 @@ func ListAppServicePlan(opts resource.ListerOpts) ([]resource.Resource, error) {
 			resources = append(resources, &AppServicePlan{
 				client: client,
 				name:   *g.Name,
-				rg:     opts.ResourceGroup,
+				rg:     l.opts.ResourceGroup,
 			})
 		}
 
@@ -62,6 +65,12 @@ func ListAppServicePlan(opts resource.ListerOpts) ([]resource.Resource, error) {
 	}
 
 	return resources, nil
+}
+
+type AppServicePlan struct {
+	client web.AppServicePlansClient
+	name   string
+	rg     string
 }
 
 func (r *AppServicePlan) Remove() error {
