@@ -4,13 +4,22 @@ import (
 	"context"
 	"fmt"
 	"github.com/aws/smithy-go/ptr"
-	"github.com/ekristen/azure-nuke/pkg/resource"
-	"github.com/ekristen/azure-nuke/pkg/types"
+	"github.com/ekristen/azure-nuke/pkg/nuke"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/resource"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/types"
 	"github.com/sirupsen/logrus"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/services/preview/resources/mgmt/2021-06-01-preview/policy"
 )
+
+func init() {
+	resource.Register(resource.Registration{
+		Name:   "PolicyDefinition",
+		Scope:  nuke.Subscription,
+		Lister: PolicyDefinitionLister{},
+	})
+}
 
 type PolicyDefinition struct {
 	client      policy.DefinitionsClient
@@ -19,19 +28,45 @@ type PolicyDefinition struct {
 	policyType  string
 }
 
-func init() {
-	resource.RegisterV2(resource.Registration{
-		Name:   "PolicyDefinition",
-		Scope:  resource.Subscription,
-		Lister: ListPolicyDefinition,
-	})
+func (r *PolicyDefinition) Filter() error {
+	if r.policyType == "BuiltIn" || r.policyType == "Static" {
+		return fmt.Errorf("cannot delete policies with type %s", r.policyType)
+	}
+	return nil
 }
 
-func ListPolicyDefinition(opts resource.ListerOpts) ([]resource.Resource, error) {
-	logrus.Tracef("subscription id: %s", opts.SubscriptionId)
+func (r *PolicyDefinition) Remove() error {
+	_, err := r.client.Delete(context.TODO(), r.name)
+	return err
+}
 
-	client := policy.NewDefinitionsClient(opts.SubscriptionId)
-	client.Authorizer = opts.Authorizers.Management
+func (r *PolicyDefinition) Properties() types.Properties {
+	properties := types.NewProperties()
+
+	properties.Set("Name", r.name)
+	properties.Set("DisplayName", r.displayName)
+	properties.Set("Type", r.policyType)
+
+	return properties
+}
+
+func (r *PolicyDefinition) String() string {
+	return r.name
+}
+
+type PolicyDefinitionLister struct {
+	opts nuke.ListerOpts
+}
+
+func (l PolicyDefinitionLister) SetOptions(opts interface{}) {
+	l.opts = opts.(nuke.ListerOpts)
+}
+
+func (l PolicyDefinitionLister) List() ([]resource.Resource, error) {
+	logrus.Tracef("subscription id: %s", l.opts.SubscriptionId)
+
+	client := policy.NewDefinitionsClient(l.opts.SubscriptionId)
+	client.Authorizer = l.opts.Authorizers.Management
 	client.RetryAttempts = 1
 	client.RetryDuration = time.Second * 2
 
@@ -64,30 +99,4 @@ func ListPolicyDefinition(opts resource.ListerOpts) ([]resource.Resource, error)
 	}
 
 	return resources, nil
-}
-
-func (r *PolicyDefinition) Filter() error {
-	if r.policyType == "BuiltIn" || r.policyType == "Static" {
-		return fmt.Errorf("cannot delete policies with type %s", r.policyType)
-	}
-	return nil
-}
-
-func (r *PolicyDefinition) Remove() error {
-	_, err := r.client.Delete(context.TODO(), r.name)
-	return err
-}
-
-func (r *PolicyDefinition) Properties() types.Properties {
-	properties := types.NewProperties()
-
-	properties.Set("Name", r.name)
-	properties.Set("DisplayName", r.displayName)
-	properties.Set("Type", r.policyType)
-
-	return properties
-}
-
-func (r *PolicyDefinition) String() string {
-	return r.name
 }
