@@ -2,13 +2,22 @@ package resources
 
 import (
 	"context"
-	"github.com/ekristen/azure-nuke/pkg/resource"
-	"github.com/ekristen/azure-nuke/pkg/types"
+	"github.com/ekristen/azure-nuke/pkg/nuke"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/resource"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/types"
 	"github.com/sirupsen/logrus"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/services/preview/security/mgmt/v3.0/security"
 )
+
+func init() {
+	resource.Register(resource.Registration{
+		Name:   "SecurityWorkspace",
+		Scope:  nuke.Subscription,
+		Lister: SecurityWorkspaceLister{},
+	})
+}
 
 type SecurityWorkspace struct {
 	client security.WorkspaceSettingsClient
@@ -16,24 +25,44 @@ type SecurityWorkspace struct {
 	scope  string
 }
 
-func init() {
-	resource.RegisterV2(resource.Registration{
-		Name:   "SecurityWorkspace",
-		Scope:  resource.Subscription,
-		Lister: ListSecurityWorkspace,
-	})
+func (r *SecurityWorkspace) Remove() error {
+	_, err := r.client.Delete(context.TODO(), r.name)
+	return err
 }
 
-func ListSecurityWorkspace(opts resource.ListerOpts) ([]resource.Resource, error) {
+func (r *SecurityWorkspace) Properties() types.Properties {
+	properties := types.NewProperties()
+
+	properties.Set("Name", r.name)
+	properties.Set("Scope", r.scope)
+
+	return properties
+}
+
+func (r *SecurityWorkspace) String() string {
+	return r.name
+}
+
+// -------------------------------------------------------------
+
+type SecurityWorkspaceLister struct {
+	opts nuke.ListerOpts
+}
+
+func (l SecurityWorkspaceLister) SetOptions(opts interface{}) {
+	l.opts = opts.(nuke.ListerOpts)
+}
+
+func (l SecurityWorkspaceLister) List() ([]resource.Resource, error) {
 	log := logrus.
 		WithField("resource", "SecurityWorkspace").
-		WithField("scope", resource.Subscription).
-		WithField("subscription", opts.SubscriptionId)
+		WithField("scope", nuke.Subscription).
+		WithField("subscription", l.opts.SubscriptionId)
 
 	log.Trace("creating client")
 
-	client := security.NewWorkspaceSettingsClient(opts.SubscriptionId)
-	client.Authorizer = opts.Authorizers.Management
+	client := security.NewWorkspaceSettingsClient(l.opts.SubscriptionId)
+	client.Authorizer = l.opts.Authorizers.Management
 	client.RetryAttempts = 1
 	client.RetryDuration = time.Second * 2
 
@@ -63,22 +92,4 @@ func ListSecurityWorkspace(opts resource.ListerOpts) ([]resource.Resource, error
 	}
 
 	return resources, nil
-}
-
-func (r *SecurityWorkspace) Remove() error {
-	_, err := r.client.Delete(context.TODO(), r.name)
-	return err
-}
-
-func (r *SecurityWorkspace) Properties() types.Properties {
-	properties := types.NewProperties()
-
-	properties.Set("Name", r.name)
-	properties.Set("Scope", r.scope)
-
-	return properties
-}
-
-func (r *SecurityWorkspace) String() string {
-	return r.name
 }
