@@ -2,67 +2,29 @@ package resources
 
 import (
 	"context"
+	"github.com/ekristen/azure-nuke/pkg/nuke"
 	"time"
 
 	"github.com/sirupsen/logrus"
 
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2021-04-01/compute"
 
-	"github.com/ekristen/azure-nuke/pkg/resource"
-	"github.com/ekristen/azure-nuke/pkg/types"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/resource"
+	"github.com/ekristen/cloud-nuke-sdk/pkg/types"
 )
+
+func init() {
+	resource.Register(resource.Registration{
+		Name:   "SSHPublicKey",
+		Lister: SSHPublicKeyLister{},
+		Scope:  nuke.ResourceGroup,
+	})
+}
 
 type SSHPublicKey struct {
 	client compute.SSHPublicKeysClient
 	name   *string
 	rg     *string
-}
-
-func init() {
-	resource.RegisterV2(resource.Registration{
-		Name:   "SSHPublicKey",
-		Lister: ListSSHPublicKey,
-		Scope:  resource.ResourceGroup,
-	})
-}
-
-func ListSSHPublicKey(opts resource.ListerOpts) ([]resource.Resource, error) {
-	logrus.Tracef("subscription id: %s", opts.SubscriptionId)
-
-	client := compute.NewSSHPublicKeysClient(opts.SubscriptionId)
-	client.Authorizer = opts.Authorizers.Management
-	client.RetryAttempts = 1
-	client.RetryDuration = time.Second * 2
-
-	resources := make([]resource.Resource, 0)
-
-	logrus.Trace("attempting to list ssh key")
-
-	ctx := context.Background()
-
-	list, err := client.ListByResourceGroup(ctx, opts.ResourceGroup)
-	if err != nil {
-		return nil, err
-	}
-
-	logrus.Trace("listing ....")
-
-	for list.NotDone() {
-		logrus.Trace("list not done")
-		for _, g := range list.Values() {
-			resources = append(resources, &SSHPublicKey{
-				client: client,
-				name:   g.Name,
-				rg:     &opts.ResourceGroup,
-			})
-		}
-
-		if err := list.NextWithContext(ctx); err != nil {
-			return nil, err
-		}
-	}
-
-	return resources, nil
 }
 
 func (r *SSHPublicKey) Remove() error {
@@ -80,4 +42,53 @@ func (r *SSHPublicKey) Properties() types.Properties {
 
 func (r *SSHPublicKey) String() string {
 	return *r.name
+}
+
+// --------------------------------------
+
+type SSHPublicKeyLister struct {
+	opts nuke.ListerOpts
+}
+
+func (l SSHPublicKeyLister) SetOptions(opts interface{}) {
+	l.opts = opts.(nuke.ListerOpts)
+}
+
+func (l SSHPublicKeyLister) List() ([]resource.Resource, error) {
+	logrus.Tracef("subscription id: %s", l.opts.SubscriptionId)
+
+	client := compute.NewSSHPublicKeysClient(l.opts.SubscriptionId)
+	client.Authorizer = l.opts.Authorizers.Management
+	client.RetryAttempts = 1
+	client.RetryDuration = time.Second * 2
+
+	resources := make([]resource.Resource, 0)
+
+	logrus.Trace("attempting to list ssh key")
+
+	ctx := context.Background()
+
+	list, err := client.ListByResourceGroup(ctx, l.opts.ResourceGroup)
+	if err != nil {
+		return nil, err
+	}
+
+	logrus.Trace("listing ....")
+
+	for list.NotDone() {
+		logrus.Trace("list not done")
+		for _, g := range list.Values() {
+			resources = append(resources, &SSHPublicKey{
+				client: client,
+				name:   g.Name,
+				rg:     &l.opts.ResourceGroup,
+			})
+		}
+
+		if err := list.NextWithContext(ctx); err != nil {
+			return nil, err
+		}
+	}
+
+	return resources, nil
 }
